[REM3-413] Properly treat the scenario where a connection is pre auth closed.
See RemoteConnection.handlePreAuthCloseRequest to fully understand the code path. The remote connection write listener needs to have shutdown writes invoked to be aware that the channel is closed, or else there is no linkage between the connection and the XNIO channels layer when the channel is closed at ServerConnectionOpenListener.Initial listener. To complete the fix, the very same remote connection write listener needs to skip the write attempt if the channel is closed, which will prevent an uneeded ClosedChannelException.

If wondering when this bug is hit, notice that the write listener can be invoked after the read listener has read -1 during the execution of XNIO NioSocketConduit.handleReady, as a result of a wake up operations call.

Signed-off-by: Flavia Rainone <frainone@redhat.com>